fix(users service): solve the problem of not running the user service
Resolved dependency issue for user repository class that we had to remove; all dependencies to this
class in all cqrs elements were removed and changed. The issue of project not compiling, which
seemed to be a typeorm issue, was fixed with the help of artificial intelligence.

BREAKING CHANGE: Removed and changed all dependencies to the user repository class in all cqrs
elements in the users service.